refactor(j-s): move CaseFilesForm.tsx to CaseFiles.tsx
